requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
torch>=1.9.0
fbprophet
matplotlib>=3.4.0
tqdm>=4.62.0
pystan>=2.19.1.1
cmdstanpy>=0.9.0
